forwarding: extract core FIB resolution into fib.rs
Pure code-motion split of forwarding/mod.rs (#5650) — the final and
largest increment. Move the per-packet FIB-resolution hot path into
forwarding/fib.rs: packet-destination parse, the table / next-table
walk (lookup_forwarding_resolution* family), per-family v4/v6 resolution
and their _inner helpers, ECMP hashing (ecmp_hash_*), route choice
(choose_v4_route/choose_v6_route, ResolvedRouteV4/V6), next-hop
selection (select_route_next_hop, tunnel_next_hop_live),
no_route_resolution, classify_metadata, canonical_route_table, and the
DEFAULT_V4_TABLE/DEFAULT_V6_TABLE/MAX_NEXT_TABLE_DEPTH constants. The
std::borrow::Cow import moves with canonical_route_table.

Functions are moved verbatim; no logic, signature, or hot-path change.
The route-selection internals (choose_*, ecmp_hash_bytes/v4/v6/flow,
tunnel_next_hop_live, v4/v6 _inner) stay private next to their callers.
select_route_next_hop and ecmp_hash_flow_seeded are promoted from
private to pub(in crate::afxdp) because the forwarding tests module
(a sibling of fib) calls them directly; all other pub(super) items
become pub(in crate::afxdp) so the mod.rs glob re-export keeps every
external call site resolving unchanged. No #[inline] function is in this
batch, so none is affected.

After this batch forwarding/mod.rs is 143 lines (from 2868): imports,
submodule declarations + glob re-exports, the small zone-pair / DNS-
reply / icmp-echo / ingress-logical-ifindex helpers, and the tests
module.

Validation: cargo build clean (0 errors, 175 pre-existing warnings).
